separete redis config
